+++ apps/laboratory/src/pages/library/ethers.tsx
@@ -1,6 +1,6 @@
+import { createWeb3Modal, defaultConfig } from '@web3modal/ethers/react'
 import { EthersTests } from '../../components/Ethers/EthersTests'
 import { AppKitButtons } from '../../components/AppKitButtons'
-import { createWeb3Modal, defaultConfig } from '@web3modal/ethers/react'
 import { ThemeStore } from '../../utils/StoreUtil'
 import { EthersConstants } from '../../utils/EthersConstants'
 import { ConstantsUtil } from '../../utils/ConstantsUtil'