+++ apps/laboratory/tests/email.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test, type BrowserContext } from '@playwright/test'
+import { expect, test, type BrowserContext, type Page } from '@playwright/test'
 import { ModalWalletPage } from './shared/pages/ModalWalletPage'
 import { Email } from './shared/utils/email'
 import { ModalWalletValidator } from './shared/validators/ModalWalletValidator'
@@ -8,6 +8,7 @@ import { SECURE_WEBSITE_URL } from './shared/constants'
 let page: ModalWalletPage
 let validator: ModalWalletValidator
 let context: BrowserContext
+let browserPage: Page
 /* eslint-enable init-declarations */
 
 // -- Setup --------------------------------------------------------------------
@@ -20,7 +21,7 @@ emailTest.describe.configure({ mode: 'serial' })
 emailTest.beforeAll(async ({ browser, library }) => {
   emailTest.setTimeout(180000)
   context = await browser.newContext()
-  const browserPage = await context.newPage()
+  browserPage = await context.newPage()
 
   page = new ModalWalletPage(browserPage, library, 'email')
   validator = new ModalWalletValidator(browserPage)
@@ -62,12 +63,14 @@ emailTest('it should reject sign', async () => {
   await validator.expectRejectedSign()
 })
 
-emailTest('it should switch network and sign', async () => {
+emailTest('it should switch network and sign', async ({ library }) => {
   let targetChain = 'Polygon'
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.switchNetwork(targetChain)
+  if (library === 'wagmi') {
+    // In wagmi, after switching network, it closes the modal
+    await page.goToSettings()
+  }
   await validator.expectSwitchedNetwork(targetChain)
   await page.closeModal()
   await page.sign()
@@ -75,10 +78,12 @@ emailTest('it should switch network and sign', async () => {
   await validator.expectAcceptedSign()
 
   targetChain = 'Ethereum'
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.switchNetwork(targetChain)
+  if (library === 'wagmi') {
+    // In wagmi, after switching network, it closes the modal
+    await page.goToSettings()
+  }
   await validator.expectSwitchedNetwork(targetChain)
   await page.closeModal()
   await page.sign()
@@ -92,9 +97,7 @@ emailTest('it should show loading on page refresh', async () => {
 })
 
 emailTest('it should disconnect correctly', async () => {
-  await page.openAccount()
-  await page.openProfileView()
-  await page.openSettings()
+  await page.goToSettings()
   await page.disconnect()
   await validator.expectDisconnected()
 })