+++ packages/wagmi/exports/react/config.ts
@@ -1,2 +1,2 @@
 // -- Universal Exports -------------------------------------------------------
-export { defaultWagmiConfig } from '../../src/utils/defaultWagmiReactConfig.js'
+export { defaultWagmiConfig } from '@web3modal/base/adapters/evm/wagmi'