+++ examples/react-ethers5/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @examples/react-ethers5
 
+## 5.1.2
+
+### Patch Changes
+
+- Added Solana Auth Provider types and schemas
+
+- Updated dependencies []:
+  - @web3modal/ethers5@5.1.2
+
+## 5.1.1
+
+### Patch Changes
+
+- Update EthProvider to v 2.15.1
+
+- Updated dependencies []:
+  - @web3modal/ethers5@5.1.1
+
 ## 5.1.0
 
 ### Minor Changes