+++ packages/cdn/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @web3modal/cdn
 
+## 5.1.2
+
+### Patch Changes
+
+- Added Solana Auth Provider types and schemas
+
+- Updated dependencies []:
+  - @web3modal/ethers@5.1.2
+  - @web3modal/ethers5@5.1.2
+  - @web3modal/wagmi@5.1.2
+
+## 5.1.1
+
+### Patch Changes
+
+- Update EthProvider to v 2.15.1
+
+- Updated dependencies []:
+  - @web3modal/ethers@5.1.1
+  - @web3modal/wagmi@5.1.1
+
 ## 5.1.0
 
 ### Minor Changes