+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web3modal/common
 
+## 5.1.2
+
+### Patch Changes
+
+- Added Solana Auth Provider types and schemas
+
+## 5.1.1
+
+### Patch Changes
+
+- Update EthProvider to v 2.15.1
+
 ## 5.1.0
 
 ### Minor Changes