+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @web3modal/wallet
 
+## 5.1.2
+
+### Patch Changes
+
+- Added Solana Auth Provider types and schemas
+
+- Updated dependencies []:
+  - @web3modal/common@5.1.2
+  - @web3modal/polyfills@5.1.2
+
+## 5.1.1
+
+### Patch Changes
+
+- Update EthProvider to v 2.15.1
+
+- Updated dependencies []:
+  - @web3modal/common@5.1.1
+  - @web3modal/polyfills@5.1.1
+
 ## 5.1.0
 
 ### Minor Changes